下载content-length优化

# Conflicts:
#	download/interfaces/kits/js/napi/download_single/src/download_base_notify.cpp
#	download/services/src/download_service_ability.cpp
#	download/services/src/download_service_task.cpp

+++ download/interfaces/innerkits/include/download_notify_interface.h
@@ -23,7 +23,7 @@ namespace OHOS::Request::Download {
 class DownloadNotifyInterface : public IRemoteBroker {
 public:
     DECLARE_INTERFACE_DESCRIPTOR(u"OHOS.Download.DownloadNotifyInterface");
-    virtual void CallBack(const std::vector<uint32_t> &params) = 0;
+    virtual void CallBack(const std::vector<int64_t> &params) = 0;
 };
 
 enum {
